Relocations fix patch 2 (#2)
* Create relocations.hpp

* Create relocations.cpp

* Update CMakeLists.txt

* Update misc.hpp

* Update misc.hpp

* Update dumper.cpp
